Use recording rule priority as the primary scheduling priority.
Previously, priorities like input and channel were added to the
recording rule priority to calculate a single priority used in
scheduling.  Now, the recording rule priority is used by itself as the
primary priority.  The other priorities are summed together and used
as a secondary priority.

What this means is all programs matching a given recording rule,
including those on less desirable inputs and channels, are scheduled
before any programs matching lower priority rules.  This is more in
line with the way most users assume and expect the scheduler to work
and still allows the ancillary priorities to influence the scheduling
of specific showings.

Note: this is essentially a reintroduction of the "ComplexPriority"
scheme.