Remove the Delete Recordings screen.
This screen was nearly functionally identical to the Watch Recordings
screen.  The primary differences between the two were the default
sorting order and whether the default SELECT action deleted or played
the currently selected recording.  The one, small, hard-coded feature
of Delete Recording to display the total disk space used by a group
has been added to Watch Recordings.  The remaining difference is the
displayed information and that is controlled by the theme.  Of the
included themes, only MythCenter-wide had specail support for Delete
Recordings.  For it, the total disk usage bar-graph has been added to
Watch Recordings.