Fix Live TV startup, particularly for analog PVR-xxx tuners.
At the start of playback, ffmpeg scans the start of the file for
stream info.  While reading packets, it uses ffio_limit() to truncate
packets during attempts to read past the size of the file.  The file
size is cached at the start of playback, and may be inordinately small
for live TV files, triggering a storm of packet truncation and failure
to start playback.  We fix this by faking a large file size for live
TV playback during the initial call to avformat_find_stream_info().