Improved cropping of the bottom 8 lines of 1088-line video.
Refs #11358.  Reverts 9d9de7b75d5354c9130526311a0c74d7172ebe6d and
cleans up the bottom 8 lines by applying a crop filter to the source.
This is tested for the XVideo output method.

The VDPAU output method does not use the standard filter mechanism,
but it works fine on 1088-line video until the h.264 parser thinks the
resolution has changed from 1080 to 1088 lines.  To compensate, we
don't acknowledge a resolution change when this happens.
(cherry picked from commit c0bc92822fe63221ad8fc5cf52a6e661b37b6f21)

Note: the hard-coding of 1088->1080 is not the best way to handle
this.  It will be cleaned up in Master.