Add test target to base makefile and the libs makefile.
This introduces a script called unittests.sh which locates all files
matching test_*.pro and then runs the tests those are expected to
generate and complains if either the tests fail or it can't find
the test executable. This script is run on both make test in the
mythtv directory and make test in the mythtv/libs directory.

The lib*/test directories have been added to the subdirs that
libs.pro compiles and a clean of the test_mythtimer.pro target has
been added. This allows a make in mythtv or mythtv/libs to compile
the unit test and a make clean to remove it. It would be nice to
have these targets work from mythtv/libs/libmythbase, but I don't
know how to do that with the current .pro file arrangement.

Note: This and the MythTimer unit test are intended as templates
for other unit tests. Any suggestions for improvement are welcome.