Improve the accuracy of durations computed by the recorder and player.
For the most part, the millisecond timestamps in the seek table match
between the recorder and "mythcommflag --rebuild".

Occasionally the player seems to miss a keyframe entry that the
recoder identifies (but this has nothing to do with duration
calculation), and more rarely the recorder misses a keyframe that the
player finds.  In addition, I have a "troublesome" ATSC channel that
sometimes throws in repeat_pict frames, and that (or maybe something
unrelated) causes some timestamps to be off by 1 millisecond.