Fixes #11358.  Use dimension and cropping information from the video.
This helps avoid hard-coding of 1088->1080 truncation and manual
rounding up to a multiple of 16.  There are still many instances in
the code base, but this is progress.

Cropping in the horizontal dimension is not yet implemented.

There is still some funny business with slightly changing aspect
ratios that needs further investigation.