Change MythDate::toString() to return an empty string for invalid inputs.
Cherry picked from master for fixes/0.26.

It's possible a "0000-00-00 00:00:00" might be somewhat acceptable in
some rare cases because the all 0s value obviously represents a
special value.  The UTC to local conversions added in 0.26, however,
now causes it to change to a less obviously special, and perhaps even
misledaing, not all 0s value.  Since MythDate::toString() was a
Myth-specific replacement for QDateTime::toString() and
QDate::toString() and those functions return an empty string for
invalid inputs, we probably should too.

Refs #11405 .