Cleanup some old/unused code from MythGenericTree and fix sorting.
Strip it back to basics, a less bloated and more managable class that offers
just the features it needs to. In this first pass I've mostly slashed
and burned the fat, there's scope for further improvements through
refactoring what remains.

Member variables have been renamed to meet coding standards, methods
need similar treatment but I'll do that as a separate commit because
of the potential to break plugins.