Remove the 'grouping' behaviour from XMLTV Parser.
This isn't within the XMLTV spec, it doesn't actually group anything
it just mangles titles and has the very real potential to do so
accidentally and without explanation to the end-user. This doesn't
affect or relate to the in-spec clumpidx handling which was added in the same
commit 9 years ago.