Remove option to disable Media Monitor.
Historical instabilities in the Media Monitor have been fixed and too
much present and future functionality depends on a working
monitor. This fixes some user frustrations caused by the monitor being
disabled, such as DVD playback or CD importation failing silently
because we cannot tell if there is a disc in the drive.