Warn about reliance on the XMLTV 'TimeOffset' setting which has been removed in 0.27
The XMLTV specification states that dates and times must be either in
GMT/UTC or specify an offset from GMT/UTC in ISO8601 format. Since the
'offset' setting has caused so much confusion over the years we're
removing it and requiring all xmltv input to follow the specs.

This change also sets the default for this setting to 'Auto' which it
should have been from the beginning.