Reduce buffer size used while scanning media.
For some reasons, when playing a MOV file by AirPlay from an iPhone fails when the buffer is too big. So add a method for specific stream allowing to temporarily reduce the size of the buffer. I suspect an issue in FFmpeg caching code.
This allows to work around it, without investigating too much the "why".

Fixes #11263