Simplify action names in the cutlist editor.
Many of the cutlist editor actions contain the frame number as one of
the components of the action string.  This is unnecessary because the
frame number is always the same, i.e. the current frame number, when
the action is handled.  These action strings are now simplified, and
several methods that took the frame number as an argument are also
simplfied since that argument is redundant.  Refs #11533.