Themed menus: Allow an item to always be set as current.
The current="active" attribute is extended to allow current="y", so
that a menu item can always be set as currently selected when entering
a menu.  The exception, as before, is when returning from a submenu,
in which case that submenu is always set as current.

Synonyms for "y" are anything starting with "y", anything starting
with "t", or "1".

Refs #11533