MythMusic: don't use cdio_paranoia for CD playback
There is problem with cdio_paranoia_read_limited() getting stuck in a loop if
the CD is ejected while playing. In theory it should give up after a set
number of retries but that doesn't seem to be working in the latest version.

I tried to lock the CD tray while playing but while the media monitor supports
locking it doesn't appear to be working :( so the easiest solution for now is
to use the non paranoia cdio routines which do detect an ejected CD properly.