MythMusic: reinstate the media handler for CD playback
The media handler will detect when a CD is inserted and depending on the
'Automatically Play CDs' setting will either show the music playback screen
and start playing the CD or the tracks will be remembered allowing the user
to manually add them to the playlist either from the Quick Playlists->From CD
menu option or I'll add a node to the playlisteditor screen so they can be
added from there as well.

Note: bumps the ABI so a re-compile is necessary.