Patch for detected dead assignment in hdhrstreamhandler.cpp
Static analysis (scan-build) detected a dead assignment to
the variable buffersize.  The usage of this variable was
deleted in commit db8f60e (ticket #6881).  There was a
modification of the initialize size in commit 2be3dd9
(ticket #10414), but since the buffersize is not actually
used, the part of the commit to set the size was not
relevant.  This patch deletes the unused variable.

Fixes #11563