Change default Xinerama aspect ratio to widescreen
and re-order the list in setup.

Note that anyone using Xinerama should still specify a proper aspect
ratio using the setting:

Monitor aspect ratio
The aspect ratio of a Xinerama display cannot be queried from the
display, so it must be specified.

in frontend Appearance settings, but since many users seem to miss this
setting after enabling Xinerama, this change updates the default to a
value that's more likely to be right, today.