Eliminate GetSignal() from new MythSystem.
The ability to query the signal that killed a program is not part of POSIX,
so we can't even depend on it on all unix systems. GetExitCode() will now
return -1 if a program was killed by a signal, -2 if it is still uncollected
and 0..255 if the program exited normally, so GetSignal() is unnecessary as
a call if all it tells us is whether a signal killed the program.