Fixed seeking issues on some DVDs where, for example, trying to jump backwards could actually cause playback to jump forwards.
dvdnav_absolute_time_search in libdvdnav (renamed from the original dvdnav_time_search but functionally the same) uses a time offset with a start sector, which will only work if all NAV packets are equally spaced.  On at least one DVD that had issues, the distance between consecutive NAV packets ranged from 96 to 457.

The existing mechanism to check whether a seek was successful and, if not, narrow in on the required time was broken when jumping backwards.

Signed-off-by: Stuart Morgan <smorgan@mythtv.org>

Fixes #11572