Re-add HLS recorder
HLS recorder got removed following the IPTV branch merge a few months ago. Re-implement it using new framework
Construct it as a derivative of existing IPTV Stream Handler.
Extend IPTVChannel so it doesn't close immediately the stream handler. Instead we mark it for deletion and only delete it 5s later, re-use it if we attempt to use it again within that delay.
This is done to speed up startup as the HLS ring buffer can take a long time to start, and the signal monitor attempts to close the stream handler causing too much delay.