Mostly revert 788bc510.  It's not needed and created a new bug.
That change was intended to fix a deadlock.  In reality, it only
worked around the deadlock and added a race condition that could cause
delayed handling of reschedule requests.  It was a series of later
changes that actually fixed the deadlocks, so this fix can be fixed by
reverting it.