Fix unit test on mac.
This add requirement for mac os 10.5 minimum.

unit tests require to run against the freshly compiled non-installed library.
To do so, it makes use of rpath functionality. However, on mac rpath behaves differently and require to modify the library install_path value and amend @rpath to it

Update external libraries to be compiled in the same fashion.