Fix for (likely) typo in checking setsockopt return code
There was a logic error due to an errant parenthesis
that was (likely) intended to be used to check the
return code from setsockopt, but instead resulted in
a value of zero being used for the length of the
operand.  This patch moves the right paren to where
it should have been.

Fixes coverity 746821
Fixes #11582