Patch to address potential integer overflow
Static analysis (coverity) detected that some of the math
in the routine getLatency was performed in 32 bit arithmetic,
even though the result was 64 bits. This could result in an
(theoretical) integer overflow. Cast to 64 bits to fix the
defect.

Note that a quick check (perhaps wrong) showed that this
method was never called, so this is probably an irrelevant
identified defect.

Fixes coverity 700584
(cherry picked from commit 499584ef7e20f97c738427fffc2657a56f1617ab)

Fixes #11595

Signed-off-by: Stuart Morgan <smorgan@mythtv.org>