Player: Enable AVSync to recover quicker after a seek or program change
The a/v timecodes can get out of step by 10's of seconds after a seek
or a livetv program change.  This can cause frames to be doubled or dropped
until the a/v streams get back in step.

This patch checks the a/v timecode difference and if larger than 1 second
disables frame dropping or doubling.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>