Check user themes directory for writability before downloading theme.
This patch will check that the user's themes directory (normally
~/.mythtv/themes on a *nix system) is writable when the user enters
the Theme Chooser screen and warn the user if it is not writable.

Another check happens at theme download time to warn the user if
the selected theme can not be downloaded and installed if the themes
directory is not writable.

Fixes #10634.