Attempt to fix double-install sometimes required in Theme Chooser.
I believe there is a race condition between closing the busy popup
dialog and sending the JumpTo to reload the theme after it is
installed. This patch handles the reload in a separate event
to allow the busy popup to close before jumping out of the
Theme Chooser screen.

References #10334.