Add the REC_STARTED_WRITING system event.
This system event is sent when the recorder starts writing the first
keyframe.  This allows a client to get better precision than the
REC_STARTED event, which is sent before other actions such as channel
changing, tuning, and waiting to see the first keyframe.  Thanks to
John Poet for direction on where in the recorder code to send the
event.

The initial intention is to facilitate simplifying the external
scripts that allow capturing subtitles from an HD-PVR, currently
documented at http://www.mythtv.org/wiki/Captions_with_HD-PVR .