Allocate sufficiently sized buffer for entire segment including the unencrypted section.
A real bug.  Per the specifications, the AES decryption
will occur only on complete 16 byte blocks, with the
remainder unencrypted.  The codes allocated a buffer
only big enough for the aes encrypted data, and when
it copied the remainder, it overflowed that buffer
(anywhere from 0 to 15 bytes).

Fixes coverity 1023889

(There is some chance coverity will not understand the fix)

Signed-off-by: Jean-Yves Avenard <jyavenard@mythtv.org>