Allow the Disk I/O Storage Group scheduler to fallback to free space
The "Balanced Disk I/O" Storage Group scheduler was indeterminate when
the weights of two filesystems were equal.  This commit makes that
decision more determinate by allowing the scheduler to fall back to using
the amount of free space on the filesystems to make the decision when
two filesystems have the same weight.

Closes #10946.

Signed-off-by: Chris Pinkham <cpinkham@mythtv.org>