Support .srt files for in-progress recordings.  Refs #11618
Allow a .srt file to be reloaded for an in-progress recording after
the last subtitle is reached.  The backend is queried at most one a
second for an updated .srt file.

Still to do:

1. Do the update check in a separate thread to avoid stuttering.

2. Currently the REC_STARTED_WRITING event is not sent when the HD-PVR
switches to a new program during live TV, so the third-party script
ends up not generating .srt files for subsequent programs in a live TV
session.

3. ccextractor buffers its output, so one needs to be at least 30
seconds behind real-time to get .srt captions.  A better solution
would be to add the required functionality to mythccextractor.