Add MythUINotificationCenter
Preliminary Notification Center manager. Will allow for an application to send informative message that will be queued and displayed to the screen in a permanent or temporary fashion.

For the time being, only handle a permanent position and uses the music miniplayer theme for ease.
First client will be the AirTunes (AirPlay) server that will display artwork and information about the song being played