Check the return value from posix_fadvise
Static analysis determined that the return value
from posix_fadvise was not checked.  Check the
return, and log if failed (only debug level
logging, since fadvise should never fail in the
first place, and if it does fail it should not
matter (it is, after all, only a hint) but if
debugging, the failure might tell one something).

Fixes coverity 746748 and 746747

Fixes #11612.

Signed-off-by: Paul Harrison <pharrison@mythtv.org>