Properly handle fd number 0 in signal manager
Visual analysis (IEAEYEBALL) detected that the
check for the file descriptor value should be
"fd >= 0" and not just "fd > 0".  While a fd of
zero is commonly stdin, do the right check just
in case.

The CLOSE macro properly performs the check
for greater than or equal to zero.  This
patch just does the same thing.

Fixes #11630.

Signed-off-by: Paul Harrison <pharrison@mythtv.org>