Fix infinite reprocessing loop for booleans in command line parser.
This fixes the immediate issue, by allowing the parser loop to
distinguish between the two scenarios, and only attempt reprocessing if
it makes sense to.

Refs #11642

I'm leaving the ticket open, as we should not be reporting an error
prematurely when we're going to attempt reprocessing. We should probably
also accept various forms of "true" and "false" for booleans, rather
than accept no values.