Fix display of notifications during video playback.
Finally nailed it after several unsuccessful attempts.

The handling of the display has been moved entirely within the OSD code.
Core problem being that with certain playback context, the current apparent screen resolution is actually of the video being played. So we need to re-adjust all positions and areas so it fits properly