Replace tmdb.py with tmdb3.py
This replaces tmdb.py with tmdb3.py as the default grabber for Movie
metadata. It also updates the database schema to force the change for
existing users. The web API tmdb.py uses will be shut down after
2013-09-15, so this is a necessary change for 0.27, and will likely need
to be done for 0.26 once tmdb3.py gets a bit wider testing.