Fix DVD playback when a still timeout is set.
The timeout should start running at the end of the cell not at the beginning.  Usually the cell only contains one frame so there's little difference but where there is video, all the parts of the code that query whether we are in a still get 'true' back right from the beginning of cell.  This can cause stuttering under some circumstances (e.g. VDPAU) because frames are not buffered.