Fix occasional crash when exiting mythtv-setup
If for whatever reason a notification got displayed in mythtv-setup; exiting would have crashed. Reason was that the ScreenStack destructor delete all its children using deleteLater(). As no eventLoop is running, the deletion never actually occurs and so the slot ScreenDeleted() never got called (which would have cleaned up m_deletedScreen). So we attempted to delete the screen using a screen stack that no longer existed resulting in a crash.