Eliminate memory leak in MythRenderOpenGL
In commit 07c54f07872c764f5f355775b27d9ee6f93daf0c
a memory leak was introduced.  This patch eliminates
the use of the pointer where the lock member object
is lost and replaces it by appropriate creation in
the initializer.  This is a "minimal" fix to the issue.

Refs #11660

Signed-off-by: Paul Harrison <pharrison@mythtv.org>