Remove dispatchNow() and all that uses it, re-implementing using thread-safe methods.
dispatchNow was marked has obsoleted years ago but was still in use. Should any listeners be in a different thread, it would have crashed (like starting playback when a preview generation was occurring)
Replaces the only two methods still using this mechanism: sendPlaybackStart and sendPlaybackEnd with their thread-safe equivalent. MythMusic will have to be adapted, changes to follow

Fixes #11671