Fix metadata fanart selection.
When artwork is processed from metadata grabber results, the artwork is
put into a QMultiMap, indexed based off the type. When pulling values
back out of a QMultiMap, they are in reverse order from how they were
inserted. Since the grabber is assumed to place the highest rated
artwork at the top of the response list, it must be pulled back out of
the bottom of the QMultiMap.

Commit c3d64c80 made this change for all artwork types by fanart. Fanart
had special handling, to allow it to select different artwork for each
season of a television show. However, the logic used to select the
artwork resulted in being indexed from the beginning of the list, rather
than indexed from the back of the list, thus pulling the lowest rated
artwork rather than the highest.

Fixes #11678
(cherry picked from commit da8ac305c3dbe955459a6a12e646b81109552836)