Make consistent use of InfoMap datatype.
There was about 3 different "map" formats. All ultimately expected to convert into InfoMap has used by MythUI.
Make their use consistent, and if it's to be used as InfoMap, then make it an InfoMap.

The ultimate test being that you should easily be able to change InfoMap from being a QHash to a QMap and it would still compile.

Also, make InfoMap argument a const, if it's used as such.