Dismiss photo being currently displayed when video is about to start.
Create the notification on demand when a photo is about to be displayed so we can discard it later.

There's still something not quite right when transitioning between photos and videos. The iPhone tends to disconnect the session. I suspect the delay we take in providing playback feedback to be the cause