don't force SI sections into a PESPacket, use the new constructor instead
A variant of this has been lightly tested and I'd like to get it into 0.27
as it should fix all kinds of issues down the line.

E.g. stuff related to the SI for DVB-CAM being rejected due to expecting
a CRC which they don't have

It also removes the false positive CRC errors from the logs.

Fixes #10495
Fixes #10496