Correct sanity check on SET_BOOKMARK protocol command
The SET_BOOKMARK backend protocol command performed a sanity check to
ensure there were four arguments (plus command) sent to it, however
after 1508085e, two of the arguments denoting a 64-bit value were merged
to a single argument, and it only used three.  This updates the sanity
check to match.

A brief search does not so any instance of this actually being used int
the code, so this does not merit updating the protocol version.

Fixes #11104