Prevent DVD bookmarks being set in a menu.
If the 'Action on exit' setting is set to 'Save position and exit', a bookmark will be created automatically.  If the TVPlayback->BACK is bound, exiting the main feature will return to the DVD menu setting a bookmark in the main feature but exiting the menu will overwrite that bookmark with a new one for the menu.  Starting playback from the bookmark would put the user back in the menu rather than the feature (although selecting 'Play' from the DVD menu will usually resume from the correct position).

This can be avoided by preventing bookmarks being set in a menu.  If a more advanced bookmark manager is implemented, this restriction could be removed.