Settings: Initialise IntegerSetting to 0 in ctor
By default a setting is initialised to a null string which, if the user
does nothing else, results in an empty string written to the database.
This can cause some routines that test for a numeric value to fail e.g.
MythArchive's DVD creator fails if MythArchiveDriveSpeed is an empty string:

    Traceback (most recent call last):
      File "/home/lvr/Projects/myth/mythinstall/share/mythtv/mytharchive/scripts/mythburn.py", line 5247, in <module>
        main()
      File "/home/lvr/Projects/myth/mythinstall/share/mythtv/mytharchive/scripts/mythburn.py", line 5143, in main
        drivespeed = int(defaultsettings.get("MythArchiveDriveSpeed", "0"))
      ValueError: invalid literal for int() with base 10: ''

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>