Fix a couple potential segfaults in MythDownloadManager.
In both GetLastModified() and download(QString,bool), we were
deleting the MythDownloadInfo* even when downloadNow() failed.  Whenever
downloadNow() fails, it sets a flag to allow downloadFinished() to
clean up as if the call was async instead of sync, so we only need to
delete our MythDownloadInfo* if the download succeeded.